requirements.txt
zlib^=1.2.13
pygments^=2.11.2
gettext^=0.21.0
jupyter^=1.0.0
pip^=22.3.1
attrs^=22.1.0
toml^=0.10.2
wheel^=0.37.1
openssl^=1.1.1s
cryptography^=38.0.1
tornado^=6.2
jinja2^=3.1.2
certifi^=2022.9.24
setuptools^=65.5.0
packaging^=21.3
pyparsing^=3.0.9
typing_extensions^=4.4.0
readline^=8.2
python-dotenv^=0.21.0
flask^=2.2.2
requests^=2.28.1
flet^=0.2.4